refactor(server): restructure standalone server into chatCompletion/{internal,external}
Reorganize the always-on server folder now that it serves two distinct surfaces:

- Rename apps/client/server/quest/ -> apps/client/server/chatCompletion/.
- internal/processRoute.ts: the internal /process route (frontend Lambda ->
  WebSocket streaming) plus its shared-secret `authorize` gate, extracted from
  server.ts.
- external/completionsRoute.ts: the public /api/ai/v2/completions route (moved).
- server.ts is now thin wiring: /health, a shared SIGTERM `track` drain callback,
  and registration of the internal + external routes.

Both routes now share one `track` callback for the in-flight drain set.

Updated the folder path in the dev command (infra/questProcessorService.ts), the
Dockerfile CMD/comment, and a questProcessor doc comment. Runtime/infra names
(QuestProcessorService, the image, the SST resource, log `service` tags) are left
unchanged - this is a folder restructure only, not a service rename.
