+++ apps/client/server/chatCompletion/external/completionsRoute.ts
