+++ apps/client/server/chatCompletion/server.test.ts
