+++ infra/questProcessorService.ts
@@ -141,7 +141,7 @@ export const questProcessorService = new sst.aws.Service('QuestProcessorService'
   // still listens on 8080 (Dockerfile ENV PORT=8080, ALB forwards 80→8080). dev.url must
   // match the local port so the frontend's dispatchQuest reaches the local server.
   dev: {
-    command: 'pnpm exec tsx server/quest/server.ts',
+    command: 'pnpm exec tsx server/chatCompletion/server.ts',
     directory: 'apps/client',
     url: 'http://localhost:8788',
   },
